Fixes for gcc-3.4.x
git-svn-id: https://ravl.svn.sourceforge.net/svnroot/ravl/main/Ravl@4561 7226f50c-d813-0410-99bf-979b288ab488

+++ Core/System/StringList.cc
@@ -25,12 +25,12 @@ namespace RavlN {
     const char *place,*eos,*lstart = 0;
     // Fill out delim table.
     for(place = rdelim;*place != 0;place++)
-      delim[*place] = true;
+      delim[(int) *place] = true;
     // Begin search of string.
     eos = &string.chars()[string.length()];
     for(place = string.chars();place != eos;) {
       // Skip spaces.
-      if(delim[*place]) {
+      if(delim[(int) *place]) {
 	place++;
 	continue;
       }
@@ -38,7 +38,7 @@ namespace RavlN {
       lstart = place;
       SizeT len = 0;
       for(;place != eos;place++) {
-	if(delim[*place])
+	if(delim[(int) *place])
 	  break;
 	len++;
       }
@@ -64,12 +64,12 @@ namespace RavlN {
     const char *place,*eos;
     // Fill out delim table.
     for(place = rdelim;*place != 0;place++)
-      delim[*place] = true;
+      delim[(int) *place] = true;
     // Begin search of string.
     eos = &string.chars()[string.length()];
     for(place = string.chars();place != eos;) {
       // Skip spaces.
-      if(delim[*place]) {
+      if(delim[(int) *place]) {
 	place++;
 	continue;
       }
@@ -77,7 +77,7 @@ namespace RavlN {
       StringC work;
       bool inQuote(false);
       for(;place != eos;place++) {
-	if(delim[*place] && !inQuote)
+	if(delim[(int) *place] && !inQuote)
 	  break;
 	switch(*place)
 	  {
